+++ src/theme/scaffold.js
@@ -1,18 +1,6 @@
 /** @module */
 import Theme from '../theme'
 
-const emojiFonts = [
-  'Apple Color Emoji',
-  'Segoe UI Emoji',
-  'Noto Color Emoji',
-  'Segoe UI Symbol',
-  'Android Emoji',
-  'Twitter Color Emoji',
-  'EmojiOne Color',
-  'Symbola',
-  'EmojiSymbols',
-]
-
 const css = `
 section {
   width: 1280px;
@@ -38,10 +26,6 @@ section:not([data-marpit-pagination])::after {
   display: none;
 }
 
-[data-marpit-emoji] {
-  font-family: ${emojiFonts.map(f => `'${f}'`).join(',')};
-}
-
 /* Normalization */
 h1 {
   font-size: 2em;
@@ -54,7 +38,6 @@ h1 {
  *
  * - Define the default slide size.
  * - Set default style for `<section>`.
- * - Set unicode emoji style for using system emoji fonts.
  * - Normalize `<h1>` heading style.
  *
  * @alias module:theme/scaffold